+++ NEWS.md
@@ -3,9 +3,15 @@
 ### brmss 0.2 (2025-10-25) (GitHub Only)
 
 * Make C++ classes 'BVS_weibull' and 'BVS_dirichlet' for every model's MCMC for-loop, likelihood, gammaSample, which makes the code more generalizable
-* First draft version at `ocbe-uio` github
+* Implement linear regression
 * TODO: Check if better techniques instead of hard upper- or lower-bounds for exponential/log scales!
 * TODO: add a complete version of proposal ratio of M-H
+* TODO: change variable name 'proposedGamma' to 'gammas_proposal'
+* TODO: add linear regression
+* TODO: add Poisson regression
+* TODO: add probit regression
+* TODO: add logistic regression (think of Pólya-gamma augmentation)
+* TODO: add multivariate linear regression (mgaussian)
 
 ### brmss 0.1 (2025-10-27) (GitHub Only)